pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "python-can-candle"
description = "CAN driver for Geschwister Schneider USB/CAN devices"
authors = [
  { name="Yu He", email="chinaheyu@outlook.com" },
]
readme = "README.md"
requires-python = ">=3.8"
license = {file = "LICENSE"}
keywords = ["gs_usb", "can", "candleLight"]
dynamic = ["version"]
dependencies = [
  "python-can >= 4.0.0",
  "candle-api == 0.0.8"
]

[project.urls]
Homepage = "https://github.com/chinaheyu/python-can-candle"
Issues = "https://github.com/chinaheyu/python-can-candle/issues"

[project.entry-points."can.interface"]
candle = "candle.candle_bus:CandleBus"

[tool.setuptools.packages.find]
where = ["."]
include = ["candle"]
namespaces = false

[tool.setuptools.dynamic]
version = {attr = "candle.__version__"}

[tool.setuptools.package-data]
candle = ["py.typed"]
